requirements.txt

# reading environment vars:
python-dotenv

# database (google sheets):
gspread==6.0.2
#google-auth==2.28.2
#google-auth-oauthlib==1.2.0

# testing:
pytest

# releasing:
# twine
